% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggests.R
\name{check_suggests}
\alias{check_suggests}
\title{Check for suggests}
\usage{
check_suggests()
}
\value{
A character vector
}
\description{
Check for suggests packages
}
\details{
Details here.
}
\author{
Fernando Mayer
}
