% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gesla.R
\name{download_gesla}
\alias{download_gesla}
\title{Download GESLA data}
\usage{
download_gesla(messages = TRUE, ask = TRUE, overwrite = FALSE)
}
\arguments{
\item{messages}{Include informative messages?}

\item{ask}{Ask for confirmation}

\item{overwrite}{Overwrite (download again)?}
}
\value{
Parquet files
}
\description{
Download GESLA dataset
}
\details{
Details of download.
}
\author{
Fernando Mayer
}
