% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gesla.R
\name{read_gesla}
\alias{read_gesla}
\title{Read a GESLA dataset}
\usage{
read_gesla(file, as_data_frame = FALSE, ...)
}
\arguments{
\item{file}{The file name (must be csv or parquet only)}

\item{as_data_frame}{If FALSE (default), the data will be imported
as an Arrow Table format. Otherwise, the data will be in a
data.frame (tibble) format. See Details.}

\item{...}{Other arguments from \code{\link[arrow:read_delim_arrow]{arrow::read_csv_arrow()}}, and
\code{\link[arrow:read_parquet]{arrow::read_parquet()}}, from the arrow package.}
}
\value{
An Arrow Table object, or a data.frame (tibble)
}
\description{
Read a csv or parquet file, as exported from the geslaR web
interface
}
\details{
The read_gesla function will automatically identify the file type
(csv or parquet) and use the appropriate functions to read an
exported Gesla dataset.
}
\author{
Fernando Mayer
}
