% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gesla_app.R
\name{run_gesla_app}
\alias{run_gesla_app}
\title{Run the Gesla app}
\usage{
run_gesla_app(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrite?}
}
\value{
A shiny interface.
}
\description{
Run the Gesla Shiny app locally.
}
\details{
Write details here.
}
\author{
Fernando Mayer
}
