% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gesla.R
\name{download_gesla}
\alias{download_gesla}
\title{Download GESLA data}
\usage{
download_gesla(
  dest = "./gesla_dataset",
  ask = TRUE,
  messages = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dest}{The destination of the dataset}

\item{ask}{Ask for confirmation}

\item{messages}{Include informative messages?}

\item{overwrite}{Overwrite (download again)?}
}
\value{
Parquet files
}
\description{
Download GESLA dataset
}
\details{
Details of download.
}
\author{
Fernando Mayer
}
