% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gesla_app.R
\name{run_gesla_app}
\alias{run_gesla_app}
\title{Run the Gesla app}
\usage{
run_gesla_app(
  app_dest = "./gesla_app",
  dest = paste0(app_dest, "/gesla_dataset"),
  overwrite = FALSE
)
}
\arguments{
\item{app_dest}{The app folder}

\item{dest}{The dataset folder}

\item{overwrite}{Overwrite?}
}
\value{
A shiny interface.
}
\description{
Run the Gesla Shiny app locally.
}
\details{
Write details here.
}
\author{
Fernando Mayer
}
